Improved the API "menu" response objects. Moved/removed old objects.

+++ Solution/Fabric.New.Api.Objects/FabResponse.cs
@@ -5,7 +5,7 @@ namespace Fabric.New.Api.Objects {
 	/*================================================================================================*/
 	public class FabResponse {
 
-		public int TotalMs { get; set; }
+		public double TotalMs { get; set; }
 		public FabError Error { get; set; }
 
 	}
